Move api mocking code to top level conftest
